+++ src/main/java/com/onemorething/layered/menu/application/dto/MemberDTO.java
@@ -8,16 +8,13 @@ public class MemberDTO {
     private String userName;
     private String userPwd;
     private String userCheckPwd;
-
-
-
-    private Date userBirth;
+    private String userBirth;
     private String userPhone;
     private String userTech1;
     private String userTech2;
     private String userTech3;
 
-    public MemberDTO(String userEmail, String userName, String userPwd, String userCheckPwd, Date userBirth, String userPhone, String userTech1, String userTech2, String userTech3) {
+    public MemberDTO(String userEmail, String userName, String userPwd, String userCheckPwd, String userBirth, String userPhone, String userTech1, String userTech2, String userTech3) {
         this.userEmail = userEmail;
         this.userName = userName;
         this.userPwd = userPwd;
@@ -28,6 +25,7 @@ public class MemberDTO {
         this.userTech2 = userTech2;
         this.userTech3 = userTech3;
     }
+
 
     public String getUserEmail() {
         return userEmail;
@@ -61,11 +59,11 @@ public class MemberDTO {
         this.userCheckPwd = userCheckPwd;
     }
 
-    public Date getUserBirth() {
+    public String getUserBirth() {
         return userBirth;
     }
 
-    public void setUserBirth(Date userBirth) {
+    public void setUserBirth(String userBirth) {
         this.userBirth = userBirth;
     }
 
@@ -101,18 +99,4 @@ public class MemberDTO {
         this.userTech3 = userTech3;
     }
 
-    @Override
-    public String toString() {
-        return "MemberDTO{" +
-                "userEmail='" + userEmail + '\'' +
-                ", userName='" + userName + '\'' +
-                ", userPwd='" + userPwd + '\'' +
-                ", userCheckPwd='" + userCheckPwd + '\'' +
-                ", userBirth=" + userBirth +
-                ", userPhone='" + userPhone + '\'' +
-                ", userTech1='" + userTech1 + '\'' +
-                ", userTech2='" + userTech2 + '\'' +
-                ", userTech3='" + userTech3 + '\'' +
-                '}';
-    }
 }
