+++ src/main/java/com/onemorething/layered/menu/application/service/SignUpService.java
@@ -1,7 +1,7 @@
 package com.onemorething.layered.menu.application.service;
 
 import com.onemorething.layered.menu.application.dto.MemberDTO;
-import com.onemorething.layered.menu.domain.service.UserSignUpService;
+import com.onemorething.layered.menu.domain.service.login.UserSignUpService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,17 +20,15 @@ public class SignUpService {
     /* 회원가입 로직 */
     public void SignUp(MemberDTO memberDTO) {
 
-        System.out.println("memberDTO = " + memberDTO);
-
-        //검증로직
-        userSignUpService.validEmail(memberDTO.getUserEmail());
+        //검증로직 호출
+        userSignUpService.signupValidEmail(memberDTO.getUserEmail());
         userSignUpService.signUpValidPwd(memberDTO.getUserPwd());
-        userSignUpService.signUpCheckPassword(memberDTO.getUserPwd() ,memberDTO.getUserCheckPwd());
-        userSignUpService.signUpValidPhone(memberDTO.getUserPhone());
+        userSignUpService.signUpCheckPwd(memberDTO.getUserPwd(), memberDTO.getUserCheckPwd());
         userSignUpService.signUpValidPhone(memberDTO.getUserPhone());
 
-        // eneity 변환
+            // eneity 변환
 
-        // repo 호출하여 저장
+            // repo 호출하여 저장
+
     }
 }
