--- src/main/java/com/onemorething/layered/menu/domain/service/login/LoginApiService.java
@@ -1,4 +1,0 @@
-package com.onemorething.layered.menu.domain.service.login;
-
-public interface LoginApiService {//외부 API 이용 용도
-}
