+++ src/main/java/com/onemorething/layered/menu/domain/service/tokenlogin/LoginApiService.java
@@ -1,0 +1,4 @@
+package com.onemorething.layered.menu.domain.service.tokenlogin;
+
+public interface LoginApiService {//외부 API 이용 용도
+}
