--- src/main/java/com/onemorething/layered/menu/infra/service/login/KakaoLoginApi.java
@@ -1,4 +1,0 @@
-package com.onemorething.layered.menu.infra.service.login;
-
-public class KakaoLoginApi {
-}
