+++ src/main/resources/templates/account/login.html
@@ -5,6 +5,11 @@
     <title>Title</title>
 </head>
 <body>
-    <h1>로그인 페이지</h1>
+<h1 align="center">Sign In</h1>
+<form align="center" action="/account/login" method="post">
+    Email : <input type="text" name="userEmail"><br>
+    Password : <input type="password" name="password"><br>
+    <input type="submit" value="sign In">
+</form>
 </body>
 </html>
