+++ src/main/resources/templates/account/signup.html
@@ -2,9 +2,33 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <title>sign Up</title>
 </head>
 <body>
-    <h1>회원가입 페이지</h1>
+<script>
+    var errorMessage = "[[${errorMessage}]]";
+    //에러메시지가 오지않을경우 초기값 ""
+    //Thymemleaf 활용하여 에러 메시지 출력
+    if (errorMessage.trim() !== "") {
+        //공백을 지웠을때 "" 가 아닐경우 (값이 있을경우)
+        alert(errorMessage);
+        //AccountController에서 redirectAttributes로 보낸 errormessage
+        // alert 메시지로 출력
+    }
+</script>
+
+<h1 align="center">Sign Up</h1>
+<form align="center" action="/account/signup" method="post">
+    Email: <input type="text" name="userEmail"><br>
+    Password: <input type="password" name="userPwd"><br>
+    Repeat password: <input type="password" name="userCheckPwd"><br>
+    Name: <input type="text" name="userName"><br>
+    Phone Number: <input type="text" name="userPhone"><br>
+    Birth: <input type="date" name="userBirth"><br>
+    Technology Stack1: <input type="text" name="userTech1" ><br>
+    Technology Stack2: <input type="text" name="userTech2"><br>
+    Technology Stack3: <input type="text" name="userTech3"><br>
+    <input type="submit" value="signup">
+</form>
 </body>
 </html>
