+++ src/test/java/com/onemorething/layered/menu/domain/service/login/SignUpCheckPwdTests.java
@@ -1,5 +1,6 @@
-package com.onemorething.layered.menu.domain.service;
+package com.onemorething.layered.menu.domain.service.login;
 
+import com.onemorething.layered.menu.domain.service.login.UserSignUpService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -7,11 +8,13 @@ import org.junit.jupiter.api.Test;
 
 class SignUpCheckPwdTests {
 
-    private UserSignUpService loginImplService;
+    private UserSignUpService userSignUpService;
+    private Exception e;
 
     @BeforeEach
     public void setUp() {
-        this.loginImplService = new UserSignUpService();
+        this.e=new Exception();
+        this.userSignUpService = new UserSignUpService();
     }
 
     @DisplayName("비밀번호 재확인 일치")
@@ -22,9 +25,9 @@ class SignUpCheckPwdTests {
         String pwd1 = "123456789";
         String pwd2 = "123456789";
         // when
-        String result = loginImplService.signUpCheckPassword(pwd1, pwd2);
+        String result = userSignUpService.signUpCheckPwd(pwd1, pwd2);
         // then
-        Assertions.assertEquals("Yes", result);
+        Assertions.assertEquals("정상", result);
     }
     @DisplayName("비밀번호 재확인 불일치")
     @Test
@@ -33,9 +36,11 @@ class SignUpCheckPwdTests {
         // given
         String pwd1 = "12345678";
         String pwd2 = "123456789";
-        // when
-        String result = loginImplService.signUpCheckPassword(pwd1, pwd2);
-        // then
-        Assertions.assertEquals("No", result);
+        // when ,then
+        e= Assertions.assertThrows(
+                IllegalArgumentException.class,
+                () -> userSignUpService.signUpCheckPwd(pwd1,pwd2)
+        );
+        System.out.println(e);
     }
 }
