+++ src/test/java/com/onemorething/layered/menu/domain/service/tokenlogin/LoginApiServiceTests.java
@@ -1,4 +1,4 @@
-package com.onemorething.layered.menu.domain.service.login;
+package com.onemorething.layered.menu.domain.service.tokenlogin;
 
 public interface LoginApiServiceTests {//외부 API 이용 용도
 }
