+++ src/test/java/com/onemorething/layered/menu/domain/service/tokenlogin/StackApiServiceTests.java
@@ -1,4 +1,4 @@
-package com.onemorething.layered.menu.domain.service.login;
+package com.onemorething.layered.menu.domain.service.tokenlogin;
 
 public interface StackApiServiceTests { //github 잔디 가저오기
 }
