--- src/test/java/com/onemorething/layered/menu/infra/service/login/GoogleLoginApiTests.java
@@ -1,4 +1,0 @@
-package com.onemorething.layered.menu.infra.service.login;
-
-public class GoogleLoginApiTests {
-}
