--- src/test/java/com/onemorething/layered/menu/infra/service/login/KakaoLoginApiTests.java
@@ -1,4 +1,0 @@
-package com.onemorething.layered.menu.infra.service.login;
-
-public class KakaoLoginApiTests {
-}
