+++ src/test/java/com/onemorething/layered/menu/infra/service/tokenlogin/KakaoLoginApiTests.java
@@ -1,0 +1,4 @@
+package com.onemorething.layered.menu.infra.service.tokenlogin;
+
+public class KakaoLoginApiTests {
+}
